Documentation and cleanup
