+++ Cargo.toml
@@ -17,8 +17,10 @@ name = "monitor"
 btleplug = "0.11.1"
 clap = { version = "4.4.8", features = ["derive"] }
 crossterm = "0.27.0"
+env_logger = "0.11.3"
 hound = "3.5.1"
 libc = "0.2.150"
+log = "0.4.21"
 rand = "0.8.5"
 ratatui = "0.24.0"
 rayon = "1.8.0"
